pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "nsd_compositionality"
description = "NSD Compositionality"
readme = "README.md"
version = "0.0.1"
requires-python = ">=3.12"
license = {file = "LICENSE"}
keywords = ["functional magnetic resonance imaging"]
authors = [
  {name = "Helena Balabin", email = "helena.balabin@kuleuven.be"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "torch>=1.9.0",
  "torchvision>=0.9.1",
  "pytorch-lightning>=1.3.8",
  "hydra-core>=1.1.0",
  "hydra-colorlog>=1.1.0",
  "hydra-optuna-sweeper>=1.1.0",
  "wandb",
  "pre-commit",
  "python-dotenv",
  "rich",
  "pytest",
  "sh",
  "scikit-learn>=1.5.0",
  "seaborn",
  "jupyterlab",
  "pudb",
  "datasets",
  "evaluate",
  "matplotlib",
  "nibabel",
  "numpy",
  "pandas",
  "pycocotools",
  "transformers",
  "loguru",
  "plotly",
  "kaleido",
  "torch",
  "torchvision",
  "nltk",
  "unidecode",
  "amrlib",
  "spacy",
  "mlflow"
]

[project.optional-dependencies]
dev = [
  "tox",
  "mypy",
  "flake8",
  "pyroma",
  "isort",
  "black"
]
